backend/pyproject.toml
[tool.poetry]
name = "physical-ai-rag-backend"
version = "1.0.0"
description = "Backend for RAG chatbot integration in Physical AI textbook"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = ">=0.104.1"
uvicorn = ">=0.24.0"
cohere = ">=5.5.3"
qdrant-client = ">=1.8.0"
python-dotenv = ">=1.0.0"
pydantic = ">=2.5.0"
google-generativeai = ">=0.3.1"
markdown = ">=3.5.1"
openai = ">=1.0.0"
beautifulsoup4 = ">=4.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"